https: support moving https cert to jffs

+++ release/src/router/httpd/httpd.c
@@ -1889,8 +1889,13 @@ QTN_RESET:
 void save_cert(void)
 {
 	if (eval("tar", "-C", "/", "-czf", "/tmp/cert.tgz", "etc/cert.pem", "etc/key.pem") == 0) {
-		if (nvram_set_file("https_crt_file", "/tmp/cert.tgz", 8192)) {
-			nvram_commit_x();
+		if (nvram_match("jffs2_on", "1") && check_if_dir_exist("/jffs/https")) {
+			system("cp /tmp/cert.tgz /jffs/https/cert.tgz");
+			nvram_set("https_crt_file", "");
+		} else {
+			if (nvram_set_file("https_crt_file", "/tmp/cert.tgz", 8192)) {
+				nvram_commit_x();
+			}
 		}
 	}
 	unlink("/tmp/cert.tgz");
@@ -1901,6 +1906,8 @@ void erase_cert(void)
 	unlink("/etc/cert.pem");
 	unlink("/etc/key.pem");
 	nvram_unset("https_crt_file");
+	if (check_if_file_exist("/jffs/https/cert.tgz"))
+		system("rm /jffs/https/cert.tgz");
 	//nvram_unset("https_crt_gen");
 	nvram_set("https_crt_gen", "0");
 }
@@ -1928,12 +1935,19 @@ void start_ssl(void)
 			ok = 0;
 			if (save) {
 				fprintf(stderr, "Save SSL certificate...\n"); // tmp test
-				if (nvram_get_file("https_crt_file", "/tmp/cert.tgz", 8192)) {
-					if (eval("tar", "-xzf", "/tmp/cert.tgz", "-C", "/", "etc/cert.pem", "etc/key.pem") == 0){
+				if (check_if_file_exist("/jffs/https/cert.tgz")) {
+					if (eval("tar", "-xzf", "/jffs/https/cert.tgz", "-C", "/", "etc/cert.pem", "etc/key.pem") == 0){
 						system("cat /etc/key.pem /etc/cert.pem > /etc/server.pem");
-						ok = 1;
+						ok =1;
 					}
-					unlink("/tmp/cert.tgz");
+				} else {
+					if (nvram_get_file("https_crt_file", "/tmp/cert.tgz", 8192)) {
+						if (eval("tar", "-xzf", "/tmp/cert.tgz", "-C", "/", "etc/cert.pem", "etc/key.pem") == 0){
+							system("cat /etc/key.pem /etc/cert.pem > /etc/server.pem");
+							unlink("/tmp/cert.tgz");
+							ok = 1;
+						}
+					}
 				}
 			}
 			if (!ok) {
@@ -1948,7 +1962,7 @@ void start_ssl(void)
 			}
 		}
 
-		if ((save) && (*nvram_safe_get("https_crt_file")) == 0) {
+		if ((save) && (*nvram_safe_get("https_crt_file") == 0) && !check_if_file_exist("/jffs/https/cert.tgz")) {
 			save_cert();
 		}
